Fixed Resizing Bug
Sometimes the boxes would stop resizing, something due to
the scrolling of the divs. So made a BigBoy div on top of
each box that handles the resizing and dragging and the normal
box class with the scrolling overflow for content.

Now it resizes and drags beautifully all THE TIME :D